+++ Plugins/Events/Events.hpp
@@ -37,6 +37,7 @@ class LevelEvents;
 class PVPEvents;
 class InputEvents;
 class MaterialChangeEvents;
+class ObjectEvents;
 
 class Events : public NWNXLib::Plugin
 {
@@ -123,6 +124,7 @@ private:
     std::unique_ptr<PVPEvents> m_PVPEvents;
     std::unique_ptr<InputEvents> m_inputEvents;
     std::unique_ptr<MaterialChangeEvents> m_matChangeEvents;
+    std::unique_ptr<ObjectEvents> m_objectEvents;
 };
 
 }